# frontend-quiz-app.Server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "frontend-quiz-app.Server",
    "version": "1.0"
  },
  "paths": {
    "/api/quiz/categories": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "operationId": "GetQuizCategories",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/QuizCategoryResponse"
              }
            }
          }
        }
      }
    },
    "/api/quiz": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "operationId": "GetQuizzesByCategory",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/QuizRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/QuizResponse"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "EQuizCategory": {
      "format": "int32",
      "enum": [
        "Html",
        "Css",
        "JavaScript",
        "Accessibility"
      ],
      "type": "string"
    },
    "QuizCategoryResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "imgUrl": {
          "type": "string"
        },
        "imgBackgroundColor": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/EQuizCategory"
        }
      }
    },
    "QuizRequest": {
      "type": "object",
      "properties": {
        "encryptKeyBase64": {
          "type": "string"
        }
      }
    },
    "QuizResponse": {
      "type": "object",
      "properties": {
        "order": {
          "format": "int32",
          "type": "integer"
        },
        "question": {
          "type": "string"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "encodedAnswer": {
          "$ref": "#/definitions/TextEncryptedData"
        },
        "category": {
          "$ref": "#/definitions/EQuizCategory"
        }
      }
    },
    "TextEncryptedData": {
      "type": "object",
      "properties": {
        "encrypted": {
          "type": "string"
        },
        "iv": {
          "type": "string"
        }
      }
    }
  }
}